Refactor previewer and improve UX for compiled
