+++ README.md
@@ -63,12 +63,7 @@ matrix(
 <details>
   <summary>Turn a prototype made in Workbench into an application</summary>
 
-<<<<<<< HEAD
-Use GNOME Builder to start a new project using the JavaScript or Vala GNOME Application template and copy paste your Workbench code.
-=======
 Use GNOME Builder to start a new project using the appropriate GNOME Application template and copy paste your Workbench code.
-
-> > > > > > > main
 
 </details>
