+++ src/window.js
@@ -24,7 +24,7 @@ import prettier_babel from "./lib/prettier-babel.js";
 import prettier_postcss from "./lib/prettier-postcss.js";
 import prettier_xml from "./lib/prettier-xml.js";
 import Library, { readDemo } from "./Library/Library.js";
-import Previewer from "./Previewer.js";
+import Previewer from "./Previewer/Previewer.js";
 import Compiler from "./Compiler.js";
 import logger from "./logger.js";
 
@@ -202,11 +202,6 @@ export default function Window({ application }) {
   }
 
   async function runCode() {
-    // Ensure code does not run if panel is not visible
-    if (!panel_code.panel.visible) {
-      return;
-    }
-
     button_run.set_sensitive(false);
 
     console.clear();
@@ -258,9 +253,9 @@ export default function Window({ application }) {
         });
       });
 
-      previewer.update();
-
       if (language === "JavaScript") {
+        previewer.update();
+
         // We have to create a new file each time
         // because gjs doesn't appear to use etag for module caching
         // ?foo=Date.now() also does not work as expected
@@ -275,8 +270,10 @@ export default function Window({ application }) {
         );
         await import(`file://${file_javascript.get_path()}`);
       } else if (language === "Vala") {
+        previewer.useExternal();
+        previewer.open();
+        previewer.update();
         compiler = compiler || Compiler(data_dir);
-        previewer.updateStyle();
         await compiler.compile(langs.vala.document.buffer.text);
       }
     } catch (err) {
@@ -288,10 +285,8 @@ export default function Window({ application }) {
       }
     }
 
-    // setTimeout(() => {
     previewer.start();
     panel_ui.start();
-    // });
 
     button_run.set_sensitive(true);
     console.scrollToEnd();
@@ -302,7 +297,10 @@ export default function Window({ application }) {
     parameter_type: null,
   });
   action_run.connect("activate", () => {
-    runCode().catch(logError);
+    // Ensure code does not run if panel is not visible
+    if (panel_code.panel.visible) {
+      runCode().catch(logError);
+    }
   });
   window.add_action(action_run);
   application.set_accels_for_action("win.run", ["<Control>Return"]);
@@ -342,8 +340,15 @@ export default function Window({ application }) {
     // in the future we may let each demo decide
     settings.set_boolean("show-console", true);
 
-    if (panel_code.language === "JavaScript") {
+    previewer.useInternal();
+
+    if (panel_code.language === "JavaScript" && panel_code.panel.visible) {
       await runCode();
+    } else {
+      previewer.start();
+      panel_ui.start();
+      panel_ui.update();
+      previewer.update();
     }
 
     languages.forEach(({ document }) => document.save());
